# [WIP] Flatten aggregation for count