--- a/src/coprocessor/dag/executor/flat_agg.rs
+++ b/src/coprocessor/dag/executor/flat_agg.rs
@@ -11,38 +11,57 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::Executor;
 use coprocessor::codec::datum;
 use coprocessor::*;
 use kvproto::coprocessor::{KeyRange, Response};
 use kvproto::kvrpcpb::IsolationLevel;
 use protobuf::{Message as PbMsg, RepeatedField};
 use storage::engine::ScanMode;
-use storage::{Iterator, Key, Snapshot, SnapshotStore, StoreScanner};
+use storage::{Key, Snapshot, SnapshotStore, StoreScanner};
 use tipb::executor::ExecType;
 use tipb::expression::ExprType;
+use tipb::schema::ColumnInfo;
 use tipb::select::{Chunk, DAGRequest, SelectResponse};
 
 pub struct FlatAggExecutor<S: Snapshot + 'static> {
-    store: SnapshotStore<S>,
     scanner: StoreScanner<S>,
+
+    col: ColumnInfo,
 }
 
 impl<S: Snapshot + 'static> FlatAggExecutor<S> {
-    pub fn new(range: &KeyRange, snap: S, start_ts: u64) -> Self {
+    pub fn new(range: &KeyRange, snap: S, start_ts: u64, col: ColumnInfo) -> Self {
         let store = SnapshotStore::new(snap, start_ts, IsolationLevel::SI, true);
         let lower_bound = Some(Key::from_raw(range.get_start()));
         let upper_bound = Some(Key::from_raw(range.get_end()));
+
         let scanner = store
-            .scanner(ScanMode::Forward, true, lower_bound, upper_bound)
+            .scanner(
+                ScanMode::Forward,
+                col.get_pk_handle(),
+                lower_bound,
+                upper_bound,
+            )
             .unwrap();
-        Self { store, scanner }
+        Self { scanner, col }
     }
+
     pub fn exec(&mut self) -> Result<Response> {
+        let col_id = self.col.get_column_id();
+        let count_pk_handle = self.col.get_pk_handle();
+
         let mut count: i64 = 0;
-        while let Some((_k, _v)) = self.scanner.next()? {
-            count += 1;
+        while let Some((_key, val)) = self.scanner.next()? {
+            if count_pk_handle {
+                count += 1;
+            } else if let Some(loc) = extract_col_location(&val, col_id)? {
+                if val[loc.offset] != datum::NIL_FLAG {
+                    count += 1;
+                }
+            }
         }
+
+        // Construct response
         let mut chunk = Chunk::new();
         chunk
             .mut_rows_data()
@@ -55,7 +74,39 @@ impl<S: Snapshot + 'static> FlatAggExecutor<S> {
         Ok(resp)
     }
 }
-/// Only support `select count(*) from ...` now.
+
+struct ColLocation {
+    pub id: i64,
+    pub offset: usize,
+    pub len: usize,
+}
+
+impl ColLocation {
+    pub fn new(id: i64, offset: usize, len: usize) -> Self {
+        Self { id, offset, len }
+    }
+}
+
+fn extract_col_location(mut data: &[u8], col_id: i64) -> Result<Option<ColLocation>> {
+    if data.is_empty() || (data.len() == 1 && data[0] == datum::NIL_FLAG) {
+        return Ok(None);
+    }
+
+    let length = data.len();
+    while data.is_empty() {
+        let id = datum::decode_datum(&mut data)?.i64();
+        let offset = length - data.len();
+        let (val, rem) = datum::split_datum(data, false)?;
+        if id == col_id {
+            return Ok(Some(ColLocation::new(id, offset, val.len())));
+        }
+        data = rem;
+    }
+
+    Ok(None)
+}
+
+/// Only support `select count(*)/count(k) from ...` now.
 pub fn build_flat_agg_from_dag<S: Snapshot + 'static>(
     req: &DAGRequest,
     ranges: &[KeyRange],
@@ -65,9 +116,11 @@ pub fn build_flat_agg_from_dag<S: Snapshot + 'static>(
     if executors.len() != 2 {
         return None;
     }
+
     if ranges.len() != 1 {
         return None;
     }
+
     let mut unique = false;
     let mut cols = vec![];
     match executors[0].get_tp() {
@@ -81,36 +134,46 @@ pub fn build_flat_agg_from_dag<S: Snapshot + 'static>(
         }
         _ => panic!("unknown first executor type {:?}", executors[0].get_tp()),
     }
+
     if cols.len() != 1 {
         return None;
     }
-    if !cols[0].get_pk_handle() {
-        return None;
-    }
+
+    //    if !cols[0].get_pk_handle() {
+    //        return None;
+    //    }
+
     if !executors[1].has_aggregation() {
         return None;
     }
+
     let agg = executors[1].get_aggregation();
     let agg_funcs = agg.get_agg_func();
     if agg_funcs.len() != 1 {
         return None;
     }
+
     if !agg.get_group_by().is_empty() {
         return None;
     }
+
     if agg_funcs[0].get_tp() != ExprType::Count {
         return None;
     }
+
     let children = agg_funcs[0].get_children();
     if children.len() != 1 {
         return None;
     }
-    if children[0].get_tp() != ExprType::Int64 {
-        return None;
-    }
+
+    //    if children[0].get_tp() != ExprType::Int64 {
+    //        return None;
+    //    }
+
     Some(FlatAggExecutor::<S>::new(
         &ranges[0],
         snap,
         req.get_start_ts(),
+        cols.into_iter().next().unwrap(),
     ))
 }