## What have you changed? (mandatory)

Currently, all aggregations are using the same framework, for some simple aggregations, the cost of the framework is expensive. For `select count(*) from t` query, actually we don't need to know the detail of each column for each row, we can simply cumulate the count. For `select count(k) from t` query, we only need to care about the column `k`, and cumulate the count if it is not nil.

This pr introduce a flatten aggregation module, count(*) and count(k) are supported.

## What are the type of the changes? (mandatory)

Improvement

## How has this PR been tested? (mandatory)

None

## Does this PR affect documentation (docs/docs-cn) update? (mandatory)

No

## Does this PR affect tidb-ansible update? (mandatory)

None

## Refer to a related PR or issue link (optional)

## Benchmark result if necessary (optional)
For `select count(*) from t`:
master
```
mysql> select count(*) from sbtest.sbtest1;
+-----------+
| count(*)  |
+-----------+
| 100000000 |
+-----------+
1 row in set (5.75 sec)

mysql> select count(*) from sbtest.sbtest1;
+-----------+
| count(*)  |
+-----------+
| 100000000 |
+-----------+
1 row in set (5.74 sec)
```

this pr 
```
mysql> select count(*) from sbtest.sbtest1;
+-----------+
| count(*)  |
+-----------+
| 100000000 |
+-----------+
1 row in set (3.61 sec)

mysql> select count(*) from sbtest.sbtest1;
+-----------+
| count(*)  |
+-----------+
| 100000000 |
+-----------+
1 row in set (3.74 sec)
```

For `select count(k) from t`:
master
```
mysql> select count(k) from sbtest.sbtest1;
+-----------+
| count(k)  |
+-----------+
| 100000000 |
+-----------+
1 row in set (11.26 sec)

mysql> select count(k) from sbtest.sbtest1;
+-----------+
| count(k)  |
+-----------+
| 100000000 |
+-----------+
1 row in set (10.99 sec)
```

this pr
```
mysql> select count(k) from sbtest.sbtest1;
+-----------+
| count(k)  |
+-----------+
| 100000000 |
+-----------+
1 row in set (7.46 sec)

mysql> select count(k) from sbtest.sbtest1;
+-----------+
| count(k)  |
+-----------+
| 100000000 |
+-----------+
1 row in set (7.50 sec)
```

## Add a few positive/negative examples (optional)