@BusyJay 2018-08-24
What if it's an expression rather than a column ref like `select count(case when xxx)`?